transactions/requirements.txt
fastapi[all]==0.111.0
SQLAlchemy==2.0.31
bcrypt==4.0.1 # jwt
passlib==1.7.4 # jwt
python-jose==3.3.0
aiofiles==24.1.0 # File upload
requests==2.32.3
python-dotenv==1.0.1 # read env files
pydantic==2.8.2
python-multipart==0.0.9
sqlalchemy[asyncio]
asyncpg==0.29.0
alembic==1.13.2
pika==1.3.2
psycopg2-binary==2.9.9